0python-demo/2wcam_ffmpeg_cv2/requirements.txt
ffmpeg-python==0.2.0
future==0.18.2
numpy==1.19.5
opencv-python==4.5.1.48
pkg-resources==0.0.0
